Save noise start and end times to output file
Noise times are saved if saveNoise = true.
Note, the number of noise samples is unrelated to the number of detection events, other than that if there are no detection events, there will be no noise events.

More thinking on reducing the number of noise samples saved may be needed. Currently, if many click candidates are found but rejected, then the number of noise samples saved may vastly exceed the number of detection events saved.