Fix: ClusterTool - Network plot fails for >20000 nodes
Network color labels were not being indexed correctly in case of subsetting in composite clusters.
Despite that fix, the network plot is incredibly slow for very large networks, so we decided to not show it for now when there are >=10000 nodes.
Also caught a script renaming error that prevented remove/recluster from working.