plotting tweaks
a couple changes to fix bugs in plotting. (1) addition to deal with detection falling directly on start time of a raw file. (2) Modify how labels plot so that if an encounter label is longer than the current LTSA window, it still pltos on the LTSA. Before, if an encounter started before the current LTSA window and ended after the end of the current window, it wouldn't plot.